Split orchestrator listener module
